+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Include "found" and "expected" values for issues messages in `LowercaseKeyword`.
 - Exclude `J` and `K` by default in `ShortIdentifier`.
+- Improve compiler directive parsing in cases where the directive name is not followed by whitespace.
 
 ### Fixed